moves `TransactionDependencies` struct to `zebra-node-services`
